Clean the core utils init file
