+++ saleor/graphql/core/tests/test_validators.py
@@ -6,10 +6,13 @@ from django.core.exceptions import ValidationError
 from django.utils import timezone
 from graphql.error import GraphQLError
 
+from ....product.models import Category
 from ..validators import (
+    clean_seo_fields,
     validate_end_is_after_start,
     validate_one_of_args_is_in_query,
     validate_price_precision,
+    validate_slug_and_generate_if_needed,
 )
 
 
@@ -87,3 +90,51 @@ def test_validate_one_of_args_is_in_query_single_arg_absent():
     with pytest.raises(GraphQLError) as error:
         validate_one_of_args_is_in_query("arg1", None) is None
     assert error.value.message == "At least one of arguments is required: 'arg1'."
+
+
+def test_clean_seo_fields():
+    title = "lady title"
+    description = "fantasy description"
+    data = {"seo": {"title": title, "description": description}}
+    clean_seo_fields(data)
+    assert data["seo_title"] == title
+    assert data["seo_description"] == description
+
+
+@pytest.mark.parametrize(
+    "cleaned_input",
+    [
+        {"slug": None, "name": "test"},
+        {"slug": "", "name": "test"},
+        {"slug": ""},
+        {"slug": None},
+    ],
+)
+def test_validate_slug_and_generate_if_needed_raises_errors(category, cleaned_input):
+    with pytest.raises(ValidationError):
+        validate_slug_and_generate_if_needed(category, "name", cleaned_input)
+
+
+@pytest.mark.parametrize(
+    "cleaned_input", [{"slug": "test-slug"}, {"slug": "test-slug", "name": "test"}]
+)
+def test_validate_slug_and_generate_if_needed_not_raises_errors(
+    category, cleaned_input
+):
+    validate_slug_and_generate_if_needed(category, "name", cleaned_input)
+
+
+@pytest.mark.parametrize(
+    "cleaned_input",
+    [
+        {"slug": None, "name": "test"},
+        {"slug": "", "name": "test"},
+        {"slug": ""},
+        {"slug": None},
+        {"slug": "test-slug"},
+        {"slug": "test-slug", "name": "test"},
+    ],
+)
+def test_validate_slug_and_generate_if_needed_generate_slug(cleaned_input):
+    category = Category(name="test")
+    validate_slug_and_generate_if_needed(category, "name", cleaned_input)
