+++ saleor/graphql/core/utils/__init__.py
@@ -1,34 +1,14 @@
 import binascii
-import mimetypes
 import os
 import secrets
-from typing import TYPE_CHECKING, Union
-from uuid import UUID
+from typing import Union
 
 import graphene
-import requests
-from django.core.exceptions import ValidationError
 from graphene import ObjectType
 from graphql.error import GraphQLError
-from PIL import Image, UnidentifiedImageError
 
-from ....core.utils import generate_unique_slug
 from ....plugins.webhook.utils import APP_ID_PREFIX
-from ....thumbnail import MIME_TYPE_TO_PIL_IDENTIFIER
-
-if TYPE_CHECKING:
-    from django.db.models import Model
-
-
-Image.init()
-
-
-def clean_seo_fields(data):
-    """Extract and assign seo fields to given dictionary."""
-    seo_fields = data.pop("seo", None)
-    if seo_fields:
-        data["seo_title"] = seo_fields.get("title")
-        data["seo_description"] = seo_fields.get("description")
+from ..validators import validate_if_int_or_uuid
 
 
 def snake_to_camel_case(name):
@@ -44,154 +24,6 @@ def str_to_enum(name):
     return name.replace(" ", "_").replace("-", "_").upper()
 
 
-def is_image_mimetype(mimetype: str) -> bool:
-    """Check if mimetype is image."""
-    if mimetype is None:
-        return False
-    return mimetype.startswith("image/")
-
-
-def is_supported_image_mimetype(mimetype: str) -> bool:
-    """Check if mimetype is a mimetype that thumbnails support."""
-    if mimetype is None:
-        return False
-    return mimetype in MIME_TYPE_TO_PIL_IDENTIFIER.keys()
-
-
-def is_image_url(url: str) -> bool:
-    """Check if file URL seems to be an image."""
-    if url.endswith(".webp"):
-        # webp is not recognized by mimetypes as image
-        # https://bugs.python.org/issue38902
-        return True
-    filetype = mimetypes.guess_type(url)[0]
-    return filetype is not None and is_image_mimetype(filetype)
-
-
-def validate_image_url(url: str, field_name: str, error_code: str) -> None:
-    """Check if remote file has content type of image.
-
-    Instead of the whole file, only the headers are fetched.
-    """
-    head = requests.head(url)
-    header = head.headers
-    content_type = header.get("content-type")
-    if content_type is None or not is_supported_image_mimetype(content_type):
-        raise ValidationError(
-            {field_name: ValidationError("Invalid file type.", code=error_code)}
-        )
-
-
-def get_filename_from_url(url: str) -> str:
-    """Prepare unique filename for file from URL to avoid overwritting."""
-    file_name = os.path.basename(url)
-    name, format = os.path.splitext(file_name)
-    hash = secrets.token_hex(nbytes=4)
-    return f"{name}_{hash}{format}"
-
-
-def validate_image_file(file, field_name, error_class) -> None:
-    """Validate if the file is an image supported by thumbnails."""
-    if not file:
-        raise ValidationError(
-            {
-                field_name: ValidationError(
-                    "File is required.", code=error_class.REQUIRED
-                )
-            }
-        )
-    if not is_supported_image_mimetype(file.content_type):
-        raise ValidationError(
-            {
-                field_name: ValidationError(
-                    "Invalid file type.", code=error_class.INVALID
-                )
-            }
-        )
-    _validate_image_format(file, field_name, error_class)
-    _validate_image_exif(file, field_name, error_class)
-
-
-def _validate_image_format(file, field_name, error_class):
-    """Validate image file format."""
-    allowed_extensions = get_allowed_extensions()
-    _file_name, format = os.path.splitext(file._name)
-    if not format:
-        raise ValidationError(
-            {
-                field_name: ValidationError(
-                    "Lack of file extension.", code=error_class.INVALID
-                )
-            }
-        )
-    elif format not in allowed_extensions:
-        raise ValidationError(
-            {
-                field_name: ValidationError(
-                    "Invalid file extension. Image file required.",
-                    code=error_class.INVALID,
-                )
-            }
-        )
-
-
-def get_allowed_extensions():
-    """Return image extension lists that are supported by thumbnails."""
-    return [
-        ext.lower()
-        for ext, file_type in Image.EXTENSION.items()
-        if file_type.upper() in MIME_TYPE_TO_PIL_IDENTIFIER.values()
-    ]
-
-
-def _validate_image_exif(file, field_name, error_class):
-    try:
-        im = Image.open(file)
-        im.getexif()
-    except (SyntaxError, TypeError, UnidentifiedImageError) as e:
-        raise ValidationError(
-            {
-                field_name: ValidationError(
-                    "Invalid file. The following error was raised during the attempt "
-                    "of opening the file and getting the exchangeable image file data:"
-                    f" {str(e)}",
-                    code=error_class.INVALID.value,
-                )
-            }
-        )
-
-
-def validate_slug_and_generate_if_needed(
-    instance: "Model",
-    slugable_field: str,
-    cleaned_input: dict,
-    slug_field_name: str = "slug",
-) -> dict:
-    """Validate slug from input and generate in create mutation if is not given."""
-
-    # update mutation - just check if slug value is not empty
-    # _state.adding is True only when it's new not saved instance.
-    if not instance._state.adding:  # type: ignore
-        validate_slug_value(cleaned_input)
-        return cleaned_input
-
-    # create mutation - generate slug if slug value is empty
-    slug = cleaned_input.get(slug_field_name)
-    if not slug and slugable_field in cleaned_input:
-        slug = generate_unique_slug(instance, cleaned_input[slugable_field])
-        cleaned_input[slug_field_name] = slug
-    return cleaned_input
-
-
-def validate_slug_value(cleaned_input, slug_field_name: str = "slug"):
-    if slug_field_name in cleaned_input:
-        slug = cleaned_input[slug_field_name]
-        if not slug:
-            raise ValidationError(
-                f"{slug_field_name.capitalize()} value cannot be blank."
-            )
-
-
 def get_duplicates_items(first_list, second_list):
     """Return items that appear on both provided lists."""
     if first_list and second_list:
@@ -202,29 +34,6 @@ def get_duplicates_items(first_list, second_list):
 def get_duplicated_values(values):
     """Return set of duplicated values."""
     return {value for value in values if values.count(value) > 1}
-
-
-def validate_required_string_field(cleaned_input, field_name: str):
-    """Strip and validate field value."""
-    field_value = cleaned_input.get(field_name)
-    field_value = field_value.strip() if field_value else ""
-    if field_value:
-        cleaned_input[field_name] = field_value
-    else:
-        raise ValidationError(f"{field_name.capitalize()} is required.")
-    return cleaned_input
-
-
-def validate_if_int_or_uuid(id):
-    result = True
-    try:
-        int(id)
-    except ValueError:
-        try:
-            UUID(id)
-        except (AttributeError, ValueError):
-            result = False
-    return result
 
 
 def from_global_id_or_error(
